--- ea-rest-api/src/main/java/miu/ea/nov/service/CourseServiceImpl.java
@@ -1,7 +1,0 @@
-package miu.ea.nov.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class CourseServiceImpl implements CourseService{
-}
