+++ ea-rest-api/src/main/java/miu/ea/nov/service/impl/CourseServiceImpl.java
@@ -1,0 +1,8 @@
+package miu.ea.nov.service.impl;
+
+import miu.ea.nov.service.CourseService;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CourseServiceImpl implements CourseService {
+}
