+++ ea-rest-api/src/main/java/miu/ea/nov/service/impl/StudentServiceImpl.java
@@ -1,12 +1,13 @@
-package miu.ea.nov.service;
+package miu.ea.nov.service.impl;
 
 
 import miu.ea.nov.payload.StudentFullDto;
 import miu.ea.nov.payload.StudentMinimalDto;
+import miu.ea.nov.service.StudentService;
 import org.springframework.stereotype.Service;
 
 @Service
-public class StudentServiceImpl implements StudentService{
+public class StudentServiceImpl implements StudentService {
     @Override
     public StudentFullDto createStudent(StudentFullDto studentDto) {
         return null;
